+++ backend/kirikiri/src/main/java/co/kirikiri/service/dto/auth/request/LoginRequest.java
@@ -1,8 +1,14 @@
 package co.kirikiri.service.dto.auth.request;
 
+import jakarta.validation.constraints.NotBlank;
+
 public record LoginRequest(
-    String identifier,
-    String password
+
+        @NotBlank(message = "아이디는 빈 값일 수 없습니다.")
+        String identifier,
+
+        @NotBlank(message = "비밀번호는 빈 값일 수 없습니다.")
+        String password
 ) {
 
 }
