+++ backend/kirikiri/src/main/java/co/kirikiri/service/dto/member/request/MemberJoinRequest.java
@@ -4,26 +4,28 @@ import co.kirikiri.service.dto.member.GenderType;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
+
 import java.time.LocalDate;
 
 public record MemberJoinRequest(
-    @NotBlank(message = "아이디는 빈 값일 수 없습니다.")
-    String identifier,
 
-    @NotBlank(message = "비밀번호는 빈 값일 수 없습니다.")
-    String password,
+        @NotBlank(message = "아이디는 빈 값일 수 없습니다.")
+        String identifier,
 
-    @NotBlank(message = "닉네임은 빈 값일 수 없습니다.")
-    String nickname,
+        @NotBlank(message = "비밀번호는 빈 값일 수 없습니다.")
+        String password,
 
-    @NotBlank(message = "전화번호는 빈 값일 수 없습니다.")
-    @Pattern(regexp = "^010-\\d{4}-\\d{4}$", message = "전화번호 형식이 맞지 않습니다.")
-    String phoneNumber,
+        @NotBlank(message = "닉네임은 빈 값일 수 없습니다.")
+        String nickname,
 
-    GenderType genderType,
+        @NotBlank(message = "전화번호는 빈 값일 수 없습니다.")
+        @Pattern(regexp = "^010-\\d{4}-\\d{4}$", message = "전화번호 형식이 맞지 않습니다.")
+        String phoneNumber,
 
-    @JsonFormat(pattern = "yyMMdd")
-    LocalDate birthday
+        GenderType genderType,
+
+        @JsonFormat(pattern = "yyMMdd")
+        LocalDate birthday
 ) {
 
 }
