+++ backend/kirikiri/src/test/java/co/kirikiri/persistence/auth/RefreshTokenRepositoryTest.java
@@ -1,7 +1,5 @@
 package co.kirikiri.persistence.auth;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
 import co.kirikiri.domain.auth.RefreshToken;
 import co.kirikiri.domain.auth.vo.EncryptedToken;
 import co.kirikiri.domain.member.Gender;
@@ -13,11 +11,14 @@ import co.kirikiri.domain.member.vo.Nickname;
 import co.kirikiri.domain.member.vo.Password;
 import co.kirikiri.persistence.RepositoryTest;
 import co.kirikiri.persistence.member.MemberRepository;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Optional;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 class RefreshTokenRepositoryTest extends RepositoryTest {
 
@@ -29,7 +30,7 @@ class RefreshTokenRepositoryTest extends RepositoryTest {
 
 
     public RefreshTokenRepositoryTest(final RefreshTokenRepository refreshTokenRepository,
-        final MemberRepository memberRepository) {
+                                      final MemberRepository memberRepository) {
         this.refreshTokenRepository = refreshTokenRepository;
         this.memberRepository = memberRepository;
     }
@@ -56,7 +57,7 @@ class RefreshTokenRepositoryTest extends RepositoryTest {
 
         //when
         final Optional<RefreshToken> result = refreshTokenRepository.findByTokenAndIsRevokedFalse(
-            encryptedToken);
+                encryptedToken);
 
         //then
         assertThat(result).isNotEmpty();
