+++ backend/kirikiri/src/test/java/co/kirikiri/persistence/member/MemberRepositoryTest.java
@@ -1,7 +1,5 @@
 package co.kirikiri.persistence.member;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
 import co.kirikiri.domain.member.Gender;
 import co.kirikiri.domain.member.Member;
 import co.kirikiri.domain.member.MemberProfile;
@@ -10,10 +8,13 @@ import co.kirikiri.domain.member.vo.Identifier;
 import co.kirikiri.domain.member.vo.Nickname;
 import co.kirikiri.domain.member.vo.Password;
 import co.kirikiri.persistence.RepositoryTest;
-import java.time.LocalDate;
-import java.util.Optional;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+
+import java.time.LocalDate;
+import java.util.Optional;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 class MemberRepositoryTest extends RepositoryTest {
