+++ backend/kirikiri/src/test/java/co/kirikiri/service/MemberServiceTest.java
@@ -1,8 +1,4 @@
-package co.kirikiri.service.member;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.given;
+package co.kirikiri.service;
 
 import co.kirikiri.domain.member.Gender;
 import co.kirikiri.domain.member.Member;
@@ -16,14 +12,19 @@ import co.kirikiri.persistence.member.MemberProfileRepository;
 import co.kirikiri.persistence.member.MemberRepository;
 import co.kirikiri.service.dto.member.GenderType;
 import co.kirikiri.service.dto.member.request.MemberJoinRequest;
-import java.time.LocalDate;
-import java.util.Optional;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+
+import java.time.LocalDate;
+import java.util.Optional;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.given;
 
 @ExtendWith(MockitoExtension.class)
 class MemberServiceTest {
@@ -39,12 +40,12 @@ class MemberServiceTest {
     void 회원가입을_한다() {
         //given
         final MemberJoinRequest request = new MemberJoinRequest("identifier1", "password1!", "nickname",
-            "010-1234-5678", GenderType.MALE, LocalDate.now());
+                "010-1234-5678", GenderType.MALE, LocalDate.now());
 
         given(memberRepository.findByIdentifier(any()))
-            .willReturn(Optional.empty());
+                .willReturn(Optional.empty());
         given(memberProfileRepository.findByNickname(any()))
-            .willReturn(Optional.empty());
+                .willReturn(Optional.empty());
 
         //when
         //then
@@ -55,40 +56,40 @@ class MemberServiceTest {
     void 회원가입_시_이미_존재하는_아이디가_존재할때_예외를_던진다() {
         //given
         final MemberJoinRequest request = new MemberJoinRequest("identifier1", "password1!", "nickname",
-            "010-1234-5678", GenderType.MALE, LocalDate.now());
+                "010-1234-5678", GenderType.MALE, LocalDate.now());
         final Identifier identifier = new Identifier("identifier1");
         final Password password = new Password("password1!");
         final Nickname nickname = new Nickname("nickname");
         final String phoneNumber = "010-1234-5678";
 
         final Member member = new Member(identifier, new EncryptedPassword(password),
-            new MemberProfile(Gender.MALE, LocalDate.now(), nickname, phoneNumber));
+                new MemberProfile(Gender.MALE, LocalDate.now(), nickname, phoneNumber));
         given(memberRepository.findByIdentifier(any()))
-            .willReturn(Optional.of(member));
+                .willReturn(Optional.of(member));
 
         //when
         //then
         assertThatThrownBy(() -> memberService.join(request))
-            .isInstanceOf(ConflictException.class);
+                .isInstanceOf(ConflictException.class);
     }
 
     @Test
     void 회원가입_시_이미_존재하는_닉네임_존재할때_예외를_던진다() {
         //given
         final MemberJoinRequest request = new MemberJoinRequest("identifier1", "password1!", "nickname",
-            "010-1234-5678", GenderType.MALE, LocalDate.now());
+                "010-1234-5678", GenderType.MALE, LocalDate.now());
         final Nickname nickname = new Nickname("nickname");
         final String phoneNumber = "010-1234-5678";
 
         final MemberProfile memberProfile = new MemberProfile(Gender.MALE, LocalDate.now(), nickname, phoneNumber);
         given(memberRepository.findByIdentifier(any()))
-            .willReturn(Optional.empty());
+                .willReturn(Optional.empty());
         given(memberProfileRepository.findByNickname(any()))
-            .willReturn(Optional.of(memberProfile));
+                .willReturn(Optional.of(memberProfile));
 
         //when
         //then
         assertThatThrownBy(() -> memberService.join(request))
-            .isInstanceOf(ConflictException.class);
+                .isInstanceOf(ConflictException.class);
     }
 }
